% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_n_fn.R
\name{est_n}
\alias{est_n}
\title{Estimate expression sensivity parameter \code{gamma}.}
\usage{
est_n(td, pure_samples, full_markers, gamma, curr_n, n_reps = 2,
  err.plot = FALSE)
}
\arguments{
\item{td}{The test data generated by \code{\link{gen_test_data}}.}

\item{pure_samples}{The pure sample indicies. List of vectors. The i-th element of list is a vector of indicies (rows of data) that are pure samples of type i.}

\item{full_markers}{User supplied markers. List of vectors. List should be same length as \code{pure_samples}. Each element of list is vector of indicies (columns of Y) that will be considered markers of that particular type. Generated by \code{\link{find_markers}}.}

\item{gamma}{Expression sensitivity parameter. Single positive numeric. Can be determined by the function \code{\link{est_gamma}}.}

\item{curr_n}{Initial estimate of how many markers to use. Can either be a single integer or a vector of integers, one for each cell type. If a single integer then all cell types use that number of markers.}

\item{n_reps}{Number of optimization iterations. Single integer.}

\item{err.plot}{Generate the plots of error by \code{n} used to select \code{n}.}
}
\value{
Selected number of markers per cell type. Vector of integer.
}
\description{
Estimate expression sensivity parameter \code{gamma}.
}
