% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phats_fn.R
\name{est_phats}
\alias{est_phats}
\title{Estimate the gene type proportions.}
\usage{
est_phats(Y, markers, baseline_ests, gamma_est)
}
\arguments{
\item{Y}{Matrix of expressions to deconvolve. Rows are observations.}

\item{markers}{List of vectors. One list entry for each type. Each vector is columns of Y that are markers for respective type.}

\item{gamma_est}{Scalar. Estimated slope.}

\item{basline_ests}{List of vectors (same as markers). One list entry for each type. Vector is estimated offset for each marker of the respective type.}
}
\value{
Estimated matrix of mixing proportions.
}
\description{
Estimate the gene type proportions.
}
