% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_thetas_fn.R
\name{est_thetas}
\alias{est_thetas}
\title{Estimate the offset terms.}
\usage{
est_thetas(Y, pure_samples, markers)
}
\arguments{
\item{Y}{Matrix of expressions to deconvolve. Rows are observations.}

\item{pure_samples}{List of vectors. List element for each type. Each vector contains pure samples (rows of Y) of respective type.}

\item{markers}{List of vectors. One list entry for each type. Each vector is columns of Y that are markers for respective type.}
}
\value{
List of vectors. Each vector is estimated estimated offsets of markers for each group, resp.
}
\description{
Estimate the offset terms.
}

