% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_markers_fn.R
\name{find_markers}
\alias{find_markers}
\title{Find marker measurements for each group.}
\usage{
find_markers(Y, pure_samples, method = "eta")
}
\arguments{
\item{Y}{the expression data set.}

\item{pure_samples}{List of pure samples for each group.}

\item{method}{Method to select markers.}
}
\value{
List with two elements. ``L'' is respective ranked markers for each group and ``V'' is the corresponding values of the ranking method (lower are better).
}
\description{
Find marker measurements for each group.
}
