% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_test_data_fn.R
\name{gen_test_data}
\alias{gen_test_data}
\title{Generate resampled test data.}
\usage{
gen_test_data(Y, pure_samples, N = 500)
}
\arguments{
\item{Y}{The known expression matrix.}

\item{pure_samples}{Thee pure sample indicies.}

\item{N}{The number of test data points to generate.}
}
\value{
A list of the generated test data and mixing proportions.
}
\description{
Generate resampled test data.
}
