% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{deconv}
\alias{deconv}
\title{Deconvolve expressions.}
\usage{
deconv(Y, pure_samples, n_choose)
}
\arguments{
\item{Y}{The expression matrix.}

\item{pure_samples}{The pure sample indicies.}

\item{n_choose}{How many markers to use.}
}
\value{
Matrix of estimated mixing proportions.
}
\description{
Deconvolve expressions.
}

