% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{deconv}
\alias{deconv}
\title{Deconvolve expressions.}
\usage{
deconv(Y, pure_samples, gamma, n_choose = NULL, full_markers = NULL,
  n_test = 500, n_test_reps = 2, marker_method = "eta", seed = NULL)
}
\arguments{
\item{Y}{The expression matrix. Data frame or matrix. Each row contains expression measurements for a particular sample.}

\item{pure_samples}{The pure sample indicies. List of vectors. The i-th element of list is a vector of indicies (rows of Y) that are pure samples of type i.}

\item{gamma}{Expression sensitivity parameter. If provided as a single positive number then that value will be used for gamma. Otherwise can pass as a string indicating data type. Current support for probe-level microarray as ``microarray-probe'', gene-level microarray as ``microarray-gene'' or rna-seq as ``rna-seq''.}

\item{n_choose}{How many markers to use. Can either be a single integer or a vector of integers, one for each cell type. If a single integer then all cell types use that number of markers. If not supplied number of markers will be adaptively chosen using \code{\link{est_n}}.}

\item{full_markers}{User supplied markers. List of vectors. List should be same length as \code{pure_samples}. Each element of list is vector of indicies (columns of Y) that will be considered markers of that particular type. If not supplied then list of markers will be determined by \code{\link{find_markers}}.}

\item{n_test}{How many observations to generate for training sample used to optiimize \code{gamma} or \code{n_choose} when they are not specified. Single integer.}

\item{n_test_reps}{How many iterations to use in optimization of \code{n_choose} when it is not speicified by user. Single integer.}

\item{marker_method}{The method used to determine which measurements are markers.}

\item{gamma_test_reps}{How many iterations to use in optimization of \code{gamma} when it is not speicified by user. Single integer.}
}
\value{
Matrix of estimated mixing proportions. One row for each sample (row) in \code{Y}. One column for each cell type.
}
\description{
Deconvolve expressions.
}
\seealso{
\code{\link{find_markers}}, \code{\link{est_n}}, \code{\link{est_gamma}}
}
